restructure day and week view into separate folders
