+++ components/error-states/NoIcalUrl.tsx
@@ -1,0 +1,8 @@
+export const NoIcalUrl = () => {
+  return (
+    <div className="p-8">
+      <h1 className="text-2xl font-semibold">No ICAL_URL is set 🤷🏻</h1>
+      <p className="pt-1 text-gray-500">Please set the ICAL_URL environment variable.</p>
+    </div>
+  )
+}
