+++ tsconfig.json
@@ -36,5 +36,5 @@
   ],
   "exclude": [
     "node_modules"
-  ],
+  ]
 }
